+++ testutils/src/main/java/com/google/android/exoplayer2/testutil/FakeVideoRenderer.java
@@ -33,7 +33,7 @@ public class FakeVideoRenderer extends FakeRenderer {
   private final VideoRendererEventListener.EventDispatcher eventDispatcher;
   private final DecoderCounters decoderCounters;
   private @MonotonicNonNull Format format;
-  private long streamOffsetUs;
+  private long startPositionUs;
   private boolean renderedFirstFrameAfterReset;
   private boolean mayRenderFirstFrameAfterEnableIfNotStarted;
   private boolean renderedFirstFrameAfterEnable;
@@ -54,9 +54,10 @@ public class FakeVideoRenderer extends FakeRenderer {
   }
 
   @Override
-  protected void onStreamChanged(Format[] formats, long offsetUs) throws ExoPlaybackException {
-    super.onStreamChanged(formats, offsetUs);
-    streamOffsetUs = offsetUs;
+  protected void onStreamChanged(Format[] formats, long startPositionUs, long offsetUs)
+      throws ExoPlaybackException {
+    super.onStreamChanged(formats, startPositionUs, offsetUs);
+    this.startPositionUs = startPositionUs;
     if (renderedFirstFrameAfterReset) {
       renderedFirstFrameAfterReset = false;
     }
@@ -101,7 +102,7 @@ public class FakeVideoRenderer extends FakeRenderer {
         !renderedFirstFrameAfterEnable
             ? (getState() == Renderer.STATE_STARTED || mayRenderFirstFrameAfterEnableIfNotStarted)
             : !renderedFirstFrameAfterReset;
-    shouldProcess |= shouldRenderFirstFrame && playbackPositionUs >= streamOffsetUs;
+    shouldProcess |= shouldRenderFirstFrame && playbackPositionUs >= startPositionUs;
     if (shouldProcess && !renderedFirstFrameAfterReset) {
       @MonotonicNonNull Format format = Assertions.checkNotNull(this.format);
       eventDispatcher.videoSizeChanged(